correct early freeing of context ptr
Using ForeignPtr in conjunction with reflection is a mistake.
All instances of the context may all go out of scope, but reification
can later create new ones. If the pointer is collected and finalized
in the mean time, then this creates a seqfault-like error (GHC does
not segfault, but returns random values).